requirements.txt
fastapi
uvicorn
markitdown[all]
python-docx
fpdf
